feat: add description for VIP 362
